feat: Add initial database migration and seed data for rooms, activities, and addresses
- Created migration file to set up database schema for Activities, Address, Rooms, RoomImages, and RoomsActivity tables.
- Inserted initial data for activities, addresses, rooms, and room images.
- Established foreign key relationships between Rooms and Address, and between Rooms and Activities.

style: Add PostCSS configuration file

style: Create CSS module for room listing UI

style: Add Tailwind CSS configuration file

+++ LumeLaht_RoomApi.Application/Dto/CreateRoomRequest.cs
@@ -11,7 +11,7 @@ namespace LumeLaht_RoomApi.Application.Dto
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public double PricePerHour { get; set; }
+        public decimal PricePerHour { get; set; }
         public string Status { get; set; }
         public Guid AddressId { get; set; }
         public AddressResponse Address { get; set; }
