+++ LumeLaht_RoomApi.Application/Dto/RoomOptionDTO.cs
@@ -12,8 +12,8 @@ namespace LumeLaht_RoomApi.Application.Dto
         public string Status { get; set; }
         public string City { get; set; }
         public string Region { get; set; }
-        public double? MinPrice { get; set; }
-        public double? MaxPrice { get; set; }
+        public decimal? MinPrice { get; set; }
+        public decimal? MaxPrice { get; set; }
         public List<Guid>? ActivitiesIds { get; set; }
     }
 }
