+++ LumeLaht_RoomApi.Application/Dto/RoomResponse.cs
@@ -6,7 +6,7 @@ namespace LumeLaht_RoomApi.Application.Dto
         public Guid RoomId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double PricePerHour { get; set; }
+        public decimal PricePerHour { get; set; }
         public string Status { get; set; }
         public AddressResponse Address { get; set; }
         public List<ActivityResponse> Activity { get; set; }
